+++ src/main/scala/scodec/codecs/implicits.scala
@@ -2,6 +2,7 @@ package scodec
 package codecs
 
 import scodec.bits._
+import shapeless.Lazy
 
 import java.util.UUID
 
@@ -22,9 +23,9 @@ trait ImplicitValues {
 
 /** Provides implicit codecs for collection types. */
 trait ImplicitCollections {
-  implicit def implicitListCodec[A](implicit ccount: ImplicitCodec[Int], ca: ImplicitCodec[A]): Codec[List[A]] = listOfN(ccount, ca)
-  implicit def implicitVectorCodec[A](implicit ccount: ImplicitCodec[Int], ca: ImplicitCodec[A]): Codec[Vector[A]] = vectorOfN(ccount, ca)
-  implicit def implicitOptionCodec[A](implicit cguard: ImplicitCodec[Boolean], ca: ImplicitCodec[A]): Codec[Option[A]] = optional(cguard, ca)
+  implicit def implicitListCodec[A](implicit ccount: Lazy[Codec[Int]], ca: Lazy[Codec[A]]): Codec[List[A]] = listOfN(ccount.value, ca.value)
+  implicit def implicitVectorCodec[A](implicit ccount: Lazy[Codec[Int]], ca: Lazy[Codec[A]]): Codec[Vector[A]] = vectorOfN(ccount.value, ca.value)
+  implicit def implicitOptionCodec[A](implicit cguard: Lazy[Codec[Boolean]], ca: Lazy[Codec[A]]): Codec[Option[A]] = optional(cguard.value, ca.value)
 }
 
 /** Provides implicit codecs for common types. */